pyproject.toml
[tool.poetry]
name = "noteappbackend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
mysql-connector-python = "^9.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
